% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreditLimitModel.R
\docType{package}
\name{CreditLimitModel}
\alias{CreditLimitModel}
\alias{CreditLimitModel-package}
\title{A package for computating customer credit limits.}
\description{
The CreditLimitModel package provides a model for predicting credit limits for customers based on their JBT ratings and prior year sales information:
}
\section{CreditLimitModel functions}{

The CreditLimitModel functions ...
  The only function you will need from \pkg{CreditLimitModel} is
  \code{\link{getCreditModel}}.
}

