% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditlimittestdata.R
\docType{data}
\name{dataset_test}
\alias{dataset_test}
\title{Credit Limit Test Case}
\format{A data frame with columns:
\describe{
 \item{CreditLimit}{Credit limit}
 \item{CreditLimitLocked}{Indicator whether the credit limit is locked for prevent updates}
 \item{JBTRating}{Credit rating provided by JBT}
 \item{SalesCurrentYr}{Net sales for the current year}
 \item{SalesLastYr}{Net sales for the last year}
 \item{PlainBandSalesLast12Mo}{Net plain band sales for the last 12 months}
 \item{DesignBandSalesLast12Mo}{Net design band sales for the last 12 months}
 \item{DiamondBandSalesLast12Mo}{Net diamond band sales for the last 12 months}
 \item{AlternativeMetalSalesLast12Mo}{Net alternative metal band sales for the last 12 months}
 \item{TotalSalesLast12Mo}{Net total sales for the last 12 months}
 \item{BuyingGroupMember}{}
 \item{NumberOfStoreLocations}{}
 \item{AvgDaysOfPayCategory}{}
 \item{AvgDaysOfPay}{}
 \item{MaxDaysPastDue}{}
 \item{MaxAmtPastDue}{}
 \item{AvgAmtPastDue}{}
 \item{ReturnedPaymentCount}{}
 \item{MaxReturnedPaymentAmt}{}
 \item{RuralUrbanContinuumCode}{}
 \item{UrbanInfluenceCode}{}
 \item{MetroIndicator}{}
 \item{TotalPopulation}{}
 \item{MedianEarnings}{}
 \item{MedianRetailTradeEarnings}{}
 \item{TotalHouseholds}{}
 \item{MedianIncomeHouseholds}{}
 \item{NumberOfJewelryStores}{}
 \item{NumberOfJewelryStoresState}{}
 \item{JewelryStoreSalesState}{}
 \item{CreditLimitCategory}{Credit limit category}
}}
\usage{
dataset_test
}
\description{
Dataset containing 340 test cases
}
\keyword{datasets}
