% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCreditModel.R
\name{getCreditModel}
\alias{getCreditModel}
\title{A Model for Determining Customer Credit Limits}
\usage{
getCreditModel()
}
\description{
This function returns the most recent trained model for determining customer credit limits.
}
\examples{
getCreditModel()
}
\keyword{credit}
